git-svn-id: svn://svn.cherokee-project.com/cherokee@132 5dc97367-97f1-0310-9951-d761b3857238

+++ cherokee/logger_w3c.c
@@ -216,7 +216,7 @@ cherokee_logger_w3c_write_error  (cherokee_logger_w3c_t *logger, cherokee_connec
 	int                len;
 	struct tm         *conn_time;
 	cherokee_buffer_t *request;
-	static int        *this_timezone = NULL;
+	static long       *this_timezone = NULL;
 
 	CHEROKEE_TEMP (tmp, 200);
 
@@ -278,7 +278,7 @@ cherokee_logger_w3c_write_access (cherokee_logger_w3c_t *logger, cherokee_connec
 	int                len;
 	struct tm         *conn_time;
 	cherokee_buffer_t *request;
-	static int        *this_timezone = NULL;
+	static long       *this_timezone = NULL;
 	CHEROKEE_TEMP (tmp, 200);
 
 	/* Read the bogonow value from the server
